refactor<> 생성자 오버로당을 이용한 예외 메타데이터 경우의 수 고려
